requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.20
aiofiles==23.2.1
jinja2==3.1.3
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
requests>=2.0.1,<3.0.0
langchain==0.1.12
langchain-community==0.0.38
langchain-core==0.1.53
tzlocal==5.2
typing-extensions>=4.5.0
starlette>=0.36.3
anyio>=4.3.0
click>=8.0.0
h11>=0.8.0
httptools>=0.6.1
python-dateutil>=2.8.2
websockets>=12.0
openai>=1.12.0
mistralai>=0.0.12
